tindri-WD.1.4.4/requirements.txt
openai
requests
numpy
pandas
pillow
opencv-python
python-dotenv
